requirements.txt
python-telegram-bot==12.8
python-dotenv==0.13.0